split test repo and test repo content classes
