+++ UnitTests/GitCommandIntegrationTests/ITestRepositoryData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace GitCommandIntegrationTests
+{
+    internal interface ITestRepositoryData : IDisposable
+    {
+        string ContentPath { get; }
+    }
+}
